Project 3/Aadi/requirements.txt
streamlit>=1.51.0
langchain>=0.3.0
langchain-google-genai>=0.8.0
langchain-community>=0.3.0
weaviate-client>=3.26.0
python-dotenv>=1.0.0

